Allow authentication on the NPM registry
This is useful when using a local proxy/cache, such as Nexus.
